refactor: move queries into own folder
